+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* Renamed the cli `input-notes` command to `notes`. Now we only export notes that were created on this client as the result of a transaction.
 * Added flags to `init` command for non-interactive environments
 * Added an option to verify note existence in the chain before importing.
 * Add new store note filter to fetch multiple notes by their id in a single query.